Refactored thread_stats to use static template and config.py
